\name{get_pathinfo}
\alias{get_pathinfo}
\title{get_pathinfo}
\usage{
get_pathinfo(total, ind1, ind2, line)
}
\description{
Prints count and time when you move from A to B in subway line
total : Total station count in line
ind1 : A subway station index in line
ind1 : B subway station index in line
get_pathinfo function is just a sub-function for another function
}
\examples{
get_pathinfo(total=40, ind1=10, ind2=15, line="2")
}
