\name{get_total_count}
\alias{get_total_count}
\title{get_total_count}
\usage{
get_total_count(dat, depart_name, depart_line_name, arrival_name, arrival_line_name)
}
\description{
get total subway station path count from data.
}
\examples{
get_total_count(dat=subway_sample, depart_name = "up_Name", depart_line_name = "up_Line", arrival_name = "down_Name", arrival_line_name = "down_Line")
}
