\name{shortestpath_0}
\alias{shortestpath_0}
\title{shortestpath_0}
\usage{
shortestpath_0(depart, depart_line, arrival, arrival_line)
}
\description{
Prints shortestpath(No transfer / time depend) for subway in Seoul.
when you can't get a path, then print 100count default.
cause this is just a sub-function for shortestpath().
}
\examples{
shortestpath_0(depart="", depart_line="", arrival="", arrival_line="")
}
