\name{shortestpath_1}
\alias{shortestpath_1}
\title{shortestpath_1}
\usage{
shortestpath_1(depart, depart_line, arrival, arrival_line)
}
\description{
Prints shortestpath(One transfer / time depend) for subway in Seoul.
when you can't get a path, then print 300count default.
cause this is just a sub-function for shortestpath().
}
\examples{
shortestpath_1(depart="", depart_line="", arrival="", arrival_line="")
}
