\name{shortestpath_2}
\alias{shortestpath_2}
\title{shortestpath_2}
\usage{
shortestpath_2(depart, depart_line, arrival, arrival_line)
}
\description{
Prints shortestpath(Two transfer / time depend) for subway in Seoul.
when you can't get a path, then print 300count default.
cause this is just a sub-function for shortestpath().
}
\examples{
shortestpath_2(depart="", depart_line="", arrival="", arrival_line="")
}
