\name{shortestpath_3}
\alias{shortestpath_3}
\title{shortestpath_3}
\usage{
shortestpath_3(depart, depart_line, arrival, arrival_line)
}
\description{
Prints shortestpath(Three transfer / time depend) for subway in Seoul.
when you can't get a path, then print error message default.
cause this is just a sub-function for shortestpath().
}
\examples{
shortestpath_3(depart="", depart_line="", arrival="", arrival_line="")
}
