\name{shortestpath_4}
\alias{shortestpath_4}
\title{shortestpath_4}
\usage{
shortestpath_4(depart, depart_line, arrival, arrival_line)
}
\description{
Prints shortestpath(Three transfer / time depend) for subway in Seoul.
this function will start in shortestpath when yo can't get path(both one, two, three transfer count)
when you can't get a path, then print error message default.
cause this is just a sub-function for shortestpath().
}
\examples{
shortestpath_4(depart="", depart_line="", arrival="", arrival_line="")
}
