\name{congestion_plot}
\alias{congestion_plot}
\title{congestion_plot}
\usage{
congestion_plot(data, center, zoom, google_key)
}
\description{
`data` structure is introduced in vignette/intro.rmd please see `intro.pdf` or `intro.rmd`
and use shortestpath_db_1812 for subway shortestpath database.
}
\examples{
congestion_plot(data, center, zoom, google_key)
}
