\name{get_pathinfo}
\alias{get_pathinfo}
\title{get_pathinfo}
\usage{
get_pathinfo(total, start, end, line)
}
\description{
Prints count and time when you move from A to B in subway line
total : Total station count in line
start : A subway station index in line
end : B subway station index in line
get_pathinfo function is just a sub-function for another function
}
\examples{
get_pathinfo(total=40, start=10, end=15, line="2")
}
