\name{shortestpath}
\alias{shortestpath}
\title{shortestpath}
\usage{
shortestpath(depart, depart_line, arrival, arrival_line)
}
\description{
Prints shortestpath(time depend) for subway in Seoul also you can get a subway list in your path.
when you can't get a path, you should consider another one.
cause these function is just consider zero, one, two, Three transfer count to find shortestpath for subway.
}
\examples{
shortestpath(depart="", depart_line="", arrival="", arrival_line="")
}
