\name{shortestpath}
\alias{shortestpath}
\title{shortestpath}
\usage{
shortestpath(departure, arrival)
}
\description{
Prints shortestpath(time depend) for subway in Seoul also you can get a subway list in your path.
when you can't get a path, you should consider another one.
cause these function is just consider zero, one, two, Three transfer count to find shortestpath for subway.
}
\examples{
shortestpath(departure="", arrival="")
}
